% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNDC.Shiny-package.R
\docType{package}
\name{rNDC.Shiny-package}
\alias{rNDC.Shiny}
\alias{rNDC.Shiny-package}
\title{rNDC.Shiny: graphical user interface to LTER-LIFE's NatureDataCube}
\description{
A Shiny app to find and get data from the NatureDataCube (and the related AgroDataCube and
GroenMonitor services), built on the functions of the \pkg{rNDC} package. Start it with \code{\link[=ndc_gui]{ndc_gui()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LTER-LIFE/rNDC-Shiny}
  \item Report bugs at \url{https://github.com/LTER-LIFE/rNDC-Shiny/issues}
}

}
\author{
\strong{Maintainer}: Bruno Marcos \email{brunoasmarcos@gmail.com}

Authors:
\itemize{
  \item Daan Lichtenberg \email{D.Lichtenberg@nioo.knaw.nl}
  \item Minke Mulder
}

}
\keyword{internal}
